Fix W560 "possible truncation at implicit conversion to type unsigned char" warnings raised by Tasking compiler.

+++ wolfcrypt/src/fe_low_mem.c
@@ -546,7 +546,7 @@ void fe_load(byte *x, word32 c)
     word32 i;
 
     for (i = 0; i < sizeof(c); i++) {
-        x[i] = c;
+        x[i] = (byte)c;
         c >>= 8;
     }
 
@@ -636,7 +636,7 @@ void lm_sub(byte* r, const byte* a, const byte* b)
     c = 218;
     for (i = 0; i + 1 < F25519_SIZE; i++) {
         c += 65280 + ((word32)a[i]) - ((word32)b[i]);
-        r[i] = c;
+        r[i] = (byte)c;
         c >>= 8;
     }
 
@@ -646,7 +646,7 @@ void lm_sub(byte* r, const byte* a, const byte* b)
 
     for (i = 0; i < F25519_SIZE; i++) {
         c += r[i];
-        r[i] = c;
+        r[i] = (byte)c;
         c >>= 8;
     }
 }
@@ -661,7 +661,7 @@ void lm_neg(byte* r, const byte* a)
     c = 218;
     for (i = 0; i + 1 < F25519_SIZE; i++) {
         c += 65280 - ((word32)a[i]);
-        r[i] = c;
+        r[i] = (byte)c;
         c >>= 8;
     }
 
@@ -671,7 +671,7 @@ void lm_neg(byte* r, const byte* a)
 
     for (i = 0; i < F25519_SIZE; i++) {
         c += r[i];
-        r[i] = c;
+        r[i] = (byte)c;
         c >>= 8;
     }
 }
@@ -693,7 +693,7 @@ void fe_mul__distinct(byte *r, const byte *a, const byte *b)
             c += ((word32)a[j]) *
                 ((word32)b[i + F25519_SIZE - j]) * 38;
 
-        r[i] = c;
+        r[i] = (byte)c;
     }
 
     r[31] &= 127;
@@ -701,7 +701,7 @@ void fe_mul__distinct(byte *r, const byte *a, const byte *b)
 
     for (i = 0; i < F25519_SIZE; i++) {
         c += r[i];
-        r[i] = c;
+        r[i] = (byte)c;
         c >>= 8;
     }
 }
@@ -724,7 +724,7 @@ void fe_mul_c(byte *r, const byte *a, word32 b)
     for (i = 0; i < F25519_SIZE; i++) {
         c >>= 8;
         c += b * ((word32)a[i]);
-        r[i] = c;
+        r[i] = (byte)c;
     }
 
     r[31] &= 127;
@@ -733,7 +733,7 @@ void fe_mul_c(byte *r, const byte *a, word32 b)
 
     for (i = 0; i < F25519_SIZE; i++) {
         c += r[i];
-        r[i] = c;
+        r[i] = (byte)c;
         c >>= 8;
     }
 }
